+++ CHANGELOG.md
@@ -11,10 +11,14 @@
 
 #### Added
 - (#174) Add CHANGELOG.md file
+- (#179) [API] Add `eventFilter` on `ListenEvent` API to get notification when an event with a specific name occurs
+         [API] Add `taskFilter` on `ListenResult` API to get notification when a result from a specific task occurs
+         [API] Add `outputFilter` on `ListenResult` API to get notification when a result returns a specific output
 - (#183) Add a `configuration` attribute in the `mesg.yml` file to accept docker configuration for your service
 - (#187) Stop all services when the daemon stops
 
 #### Removed
 
 #### Fixed
+- (#179) [Doc] Outdated documentation for the CLI
 - (#185) Fix logs with extra characters when `mesg-core daemon logs`